refactor - move functions from lib.rs into toolkit.rs
